[REF] l10n_tr_nilvera_einvoice_*: merge einvoice extended module
This commit merges `l10n_tr_nilvera_einvoice_extended` module
into `l10n_tr_nilvera_einvoice`.

task-5236310
