+++ addons/l10n_tr_nilvera_einvoice/data/account_incoterms_data.xml
