+++ addons/l10n_tr_nilvera_einvoice/data/template/l10n_tr_nilvera_einvoice.account.tax.code-tr.csv
@@ -1,0 +1,39 @@
+"id","name","name@tr_TR","code","percentage","code_type"
+"l10n_tr_nilvera_einvoice.account_tax_code_212","17/4-i Services Provided in Free Zones","17/4-ı Serbest Bölgelerde Verilen Hizmetler","212","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_228","Deliveries Made Free of Charge to the Institutions and Organizations Listed in Article (17/1) of Law 17/2-b","17/2-b Kanunun (17/1) Maddesinde Sayılan Kurum ve Kuruluşlara Bedelsiz Olarak Yapılan Teslimler","228","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_250","Partial Others","Kısmi İstisna Diğerleri","250","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_301","11/1-a Export of Goods","Kısmi İstisna Diğerleri","301","0.0","export_exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_302","11/1-a Service Export","11/1-a Hizmet İhracatı","302","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_308","13/d Delivery of Incentive Investment Goods","13/d Teşvikli Yatırım Mallarının Teslimi","308","0.0","export_exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_316","11/1-a Contract Services for Customers in Free Zones","11/1-a Serbest Bölgelerdeki Müşteriler İçin Yapılan Fason Hizmetler","316","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_328","13/i Residential or Workplace Deliveries","13/i Konut veya İşyeri Teslimleri","328","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_338","Exports of Goods by Manufacturers","İmalatçıların Mal İhracatları","338","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_350","Other","İstisna Diğer","350","0.0","exception"
+"l10n_tr_nilvera_einvoice.account_tax_code_601","Construction Works and Engineering-Architecture and Survey-Project Services Performed Along with These Works","Yapım İşleri İle Bu İşlerle Birlikte İfa Edilen Mühendislik-Mimarlık ve Etüt-Proje Hizmetleri","601","0.4","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_602","Survey, Plan-Project, Consultancy, Audit and Similar Services","Etüt, Plan-Proje, Danışmanlık, Denetim ve Benzeri Hizmetler","602","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_603","Renovation, Maintenance and Repair Services for Machinery, Equipment, Fixed Assets and Vehicles","Makine, Teçhizat, Demirbaş ve Taşıtlara Ait Tadil, Bakım, Onarım Hizmetleri","603","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_604","Food Service","Yemek Servis Hizmeti","604","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_605","Organization Service","Organizasyon Hizmeti","605","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_606","Labor Supply Services","İşgücü Temin Hizmetleri","606","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_607","Private Security Service","Özel Güvenlik Hizmeti","607","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_608","Building Inspection Services","Yapı Denetim Hizmetleri","608","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_609","Textile and Garment Works, Bag and Shoe Sewing Works and Intermediary Services for These Works","Fason Olarak Yaptırılan Tekstil ve Konfeksiyon İşleri, Çanta ve Ayakkabı Dikim İşleri ve Bu İşlere Aracılık Hizmetleri","609","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_610","Customer Finding / Taking Services to Tourist Shops","Turistik Mağazalara Verilen Müşteri Bulma / Götürme Hizmetleri","610","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_611","Transactions Subject to Broadcasting, Advertising and Naming Rights Revenues of Sports Clubs","Spor Kulüplerinin Yayın, Reklam ve İsim Hakkı Gelirlerine Konu İşlemleri","611","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_612","Cleaning Service","Temizlik Hizmeti","612","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_613","Environmental and Garden Maintenance Services","Çevre ve Bahçe Bakım Hizmetleri","613","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_614","Shuttle Transportation Service","Servis Taşımacılığı Hizmeti","614","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_615","All Kinds of Printing and Printing Services","Her Türlü Baskı ve Basım Hizmetleri","615","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_616","Other Services","Diğer Hizmetler","616","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_617","Ingot Deliveries Obtained from Scrap Metal","Hurda Metalden Elde Edilen Külçe Teslimleri","617","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_618","Delivery of Copper, Zinc, Iron-Steel, Aluminum and Lead Ingots Other Than Those Obtained from Scrap Metal","Hurda Metalden Elde Edilenler Dışındaki Bakır, Çinko, Demir-Çelik, Alüminyum ve Kurşun Külçe Teslimi","618","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_619","Delivery of Copper, Zinc and Aluminum Products","Bakır, Çinko ve Alüminyum Ürünlerinin Teslimi","619","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_620","Scrap and Waste Delivery of Those Who Waive the Exception","İstisnadan Vazgeçenlerin Hurda ve Atık Teslimi","620","0.7","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_621","Delivery of Raw Materials Obtained from Metal, Plastic, Rubber, Paper and Glass Scrap and Waste","Metal, Plastik, Lastik, Kauçuk, Kağıt ve Cam Hurda ve Atıklardan Elde Edilen Hammadde Teslimi","621","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_622","Cotton, Mohair, Wool and Fleece, Raw Hide and Skin Deliveries","Pamuk, Tiftik, Yün ve Yapağı İle Ham Post ve Deri Teslimleri","622","0.9","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_623","Delivery of Wood and Forest Products","Ağaç ve Orman Ürünleri Teslimi","623","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_624","Freight Transport Service","Yük Taşımacılığı Hizmeti","624","0.2","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_625","Commercial Advertising Services","Ticari Reklam Hizmetleri","625","0.3","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_626","Other Submissions","Diğer Teslimler","626","0.2","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_627","Delivery of Iron and Steel Products","Demir-Çelik Ürünlerinin Teslimi","627","0.5","withholding"
+"l10n_tr_nilvera_einvoice.account_tax_code_701","Export Registered Sales within the scope of Article 11/1-c of the VAT Law No. 3065","3065 s. KDV Kanununun 11/1-c md. Kapsamındaki İhraç Kayıtlı Satış","701","0.0","export_registration"
