+++ addons/l10n_tr_nilvera_einvoice/demo/demo_company.xml
@@ -20,13 +20,16 @@
 
     <record id="base.partner_demo_company_tr" model="res.partner">
         <field eval="[Command.set([ref('partner_demo_company_tr_tag')])]" name="category_id"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_1045"></field>
     </record>
 
     <record id="l10n_tr.l10n_tr_res_partner_1" model="res.partner">
         <field eval="[Command.set([ref('l10n_tr_res_partner_1_tag')])]" name="category_id"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_763"></field>
     </record>
 
     <record id="l10n_tr.l10n_tr_res_partner_2" model="res.partner">
         <field eval="[Command.set([ref('l10n_tr_res_partner_2_tag')])]" name="category_id"/>
+        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice.tax_office_545"></field>
     </record>
 </odoo>
