+++ addons/l10n_tr_nilvera_einvoice/models/__init__.py
@@ -1,5 +1,15 @@
 from . import account_edi_xml_ubl_tr
 from . import account_journal
 from . import account_move
+from . import account_move_line
 from . import account_move_send
+from . import account_tax
+from . import l10n_tr_nilvera_einvoice_account_tax_code
+from . import l10n_tr_nilvera_einvoice_tax_office
+from . import product_product
+from . import product_template
+from . import res_company
+from . import res_config_settings
+from . import res_partner
 from . import res_partner_category
+from . import template_tr
