+++ addons/l10n_tr_nilvera_einvoice/models/account_move.py
@@ -49,6 +49,80 @@ class AccountMove(models.Model):
         copy=False,
         default='not_sent',
     )
+    l10n_tr_gib_invoice_scenario = fields.Selection(
+        selection=[
+            ('TEMELFATURA', "Basic"),
+            ('KAMU', "Public Sector"),
+        ],
+        default='TEMELFATURA',
+        string="Invoice Scenario",
+        help="The scenario of the invoice to be sent to GİB.",
+    )
+    l10n_tr_gib_invoice_type = fields.Selection(
+        compute='_compute_l10n_tr_gib_invoice_type',
+        store=True,
+        readonly=False,
+        default='SATIS',
+        string="GIB Invoice Type",
+        selection=[
+            ('SATIS', "Sales"),
+            ('TEVKIFAT', "Withholding"),
+            ('IHRACKAYITLI', "Registered for Export"),
+            ('ISTISNA', "Tax Exempt"),
+        ],
+        help="The type of invoice to be sent to GİB.",
+    )
+    l10n_tr_is_export_invoice = fields.Boolean(string="GİB Product Export Invoice")
+    l10n_tr_shipping_type = fields.Selection(
+        selection=[
+            ('1', "Sea Transportation"),
+            ('2', "Railway Transportation"),
+            ('3', "Road Transportation"),
+            ('4', "Air Transportation"),
+            ('5', "Post"),
+            ('6', "Combined Transportation"),
+            ('7', "Fixed Transportation"),
+            ('8', "Domestic Water Transportation"),
+            ('9', "Invalid Transportation Method"),
+        ],
+        string="Shipping Method",
+        help="The type of shipping.",
+    )
+    l10n_tr_exemption_code_id = fields.Many2one(
+        comodel_name='l10n_tr_nilvera_einvoice.account.tax.code',
+        compute='_compute_l10n_tr_exemption_code_id',
+        store=True,
+        readonly=False,
+        string="Exemption Reason",
+        help="The exception reason of the invoice.",
+    )
+    l10n_tr_exemption_code_domain_list = fields.Binary(compute='_compute_l10n_tr_exemption_code_domain_list')
+    l10n_tr_nilvera_customer_status = fields.Selection(
+        string="Partner Nilvera Status",
+        related='partner_id.l10n_tr_nilvera_customer_status',
+    )
+
+    @api.depends("l10n_tr_gib_invoice_scenario", "l10n_tr_gib_invoice_type", "l10n_tr_is_export_invoice")
+    def _compute_l10n_tr_exemption_code_domain_list(self):
+        for record in self:
+            domain = []
+            if record.l10n_tr_gib_invoice_type == "ISTISNA":
+                domain.extend(("exception", "export_exception"))
+            if record.l10n_tr_gib_invoice_type == "IHRACKAYITLI":
+                domain.append("export_registration")
+            if record.l10n_tr_is_export_invoice:
+                domain.append("export_exception")
+            record.l10n_tr_exemption_code_domain_list = domain
+
+    @api.depends("l10n_tr_gib_invoice_scenario", "l10n_tr_is_export_invoice")
+    def _compute_l10n_tr_gib_invoice_type(self):
+        for record in self:
+            record.l10n_tr_gib_invoice_type = False
+
+    @api.depends("l10n_tr_gib_invoice_scenario", "l10n_tr_gib_invoice_type", "partner_id")
+    def _compute_l10n_tr_exemption_code_id(self):
+        for record in self:
+            record.l10n_tr_exemption_code_id = False
 
     def _get_import_file_type(self, file_data):
         """ Identify Nilvera UBL files. """
@@ -72,10 +146,8 @@ class AccountMove(models.Model):
         return CATEGORY_MOVE_TYPE_MAP.get(document_category.lower())
 
     def _l10n_tr_get_status_invoice_channel(self):
-        # Overriden in l10n_tr_nilvera_einvoice_extended
-        # to handle Export invoices edge case
         self.ensure_one()
-        return self.partner_id.l10n_tr_nilvera_customer_status
+        return 'einvoice' if self.l10n_tr_is_export_invoice else self.partner_id.l10n_tr_nilvera_customer_status
 
     @api.model
     def _get_ubl_cii_builder_from_xml_tree(self, tree):
@@ -358,7 +430,11 @@ class AccountMove(models.Model):
     def _get_partner_l10n_tr_nilvera_customer_alias_name(self):
         # Allows overriding the default customer alias with a custom one.
         self.ensure_one()
-        return self.partner_id.l10n_tr_nilvera_customer_alias_id.name
+        return (
+            self.l10n_tr_is_export_invoice
+            and self.company_id.l10n_tr_nilvera_export_alias
+            or self.partner_id.l10n_tr_nilvera_customer_alias_id.name
+        )
 
     # -------------------------------------------------------------------------
     # CRONS
