+++ addons/l10n_tr_nilvera_einvoice/models/l10n_tr_nilvera_einvoice_account_tax_code.py
@@ -1,8 +1,8 @@
 from odoo import api, fields, models
 
 
-class L10nTrNilveraEinvoiceExtendedAccountTaxCode(models.Model):
-    _name = "l10n_tr_nilvera_einvoice_extended.account.tax.code"
+class L10nTrNilveraEinvoiceAccountTaxCode(models.Model):
+    _name = "l10n_tr_nilvera_einvoice.account.tax.code"
     _description = "Turkish Tax Codes (GIB Codes)"
 
     name = fields.Char(string="Reason", required=True, translate=True)
