+++ addons/l10n_tr_nilvera_einvoice/models/template_tr.py
@@ -7,10 +7,10 @@ class AccountChartTemplate(models.AbstractModel):
 
     @template("tr", "account.tax")
     def _get_tr_withholding_account_tax(self):
-        additional = self._parse_csv("tr", "account.tax", module="l10n_tr_nilvera_einvoice_extended")
+        additional = self._parse_csv("tr", "account.tax", module="l10n_tr_nilvera_einvoice")
         self._deref_account_tags("tr", additional)
         return additional
 
-    @template("tr", "l10n_tr_nilvera_einvoice_extended.account.tax.code")
+    @template("tr", "l10n_tr_nilvera_einvoice.account.tax.code")
     def _get_tr_withholding_account_tax_code(self):
-        return self._parse_csv("tr", "l10n_tr_nilvera_einvoice_extended.account.tax.code", module="l10n_tr_nilvera_einvoice_extended")
+        return self._parse_csv("tr", "l10n_tr_nilvera_einvoice.account.tax.code", module="l10n_tr_nilvera_einvoice")
