+++ addons/l10n_tr_nilvera_einvoice/tests/__init__.py
@@ -1,3 +1,4 @@
 
 from . import test_account_move_send
+from . import test_xml_ubl_tr
 from . import test_xml_ubl_tr_common
