+++ addons/l10n_tr_nilvera_einvoice/tests/expected_xmls/invoice_basic_sale_earchive.xml
