+++ addons/l10n_tr_nilvera_einvoice/tests/expected_xmls/invoice_einvoice_multicurrency.xml
