+++ addons/l10n_tr_nilvera_einvoice/views/account_tax_views.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="view_account_tax_form_view_l10n_tr_nilvera_extended" model="ir.ui.view">
-        <field name="name">account.tax.form.view.inherit.l10n_tr_nilvera_extended</field>
+    <record id="view_account_tax_form_view_l10n_tr_nilvera" model="ir.ui.view">
+        <field name="name">account.tax.form.view.inherit.l10n_tr_nilvera</field>
         <field name="model">account.tax</field>
         <field name="inherit_id" ref="account.view_tax_form"/>
         <field name="arch" type="xml">
