+++ addons/l10n_tr_nilvera_einvoice/views/l10n_tr_nilvera_einvoice_account_tax_code_views.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="l10n_tr_nilvera_einvoice_extended_account_tax_code_view_list" model="ir.ui.view">
-        <field name="name">l10n_tr_nilvera_einvoice_extended.account.tax.code.list</field>
-        <field name="model">l10n_tr_nilvera_einvoice_extended.account.tax.code</field>
+    <record id="l10n_tr_nilvera_einvoice_account_tax_code_view_list" model="ir.ui.view">
+        <field name="name">l10n_tr_nilvera_einvoice.account.tax.code.list</field>
+        <field name="model">l10n_tr_nilvera_einvoice.account.tax.code</field>
         <field name="arch" type="xml">
             <list string="Tax Codes" multi_edit="1">
                 <field name="name"/>
@@ -13,9 +13,9 @@
         </field>
     </record>
 
-    <record id="l10n_tr_nilvera_einvoice_extended_account_tax_code_view_form" model="ir.ui.view">
-        <field name="name">l10n_tr_nilvera_einvoice_extended.account.tax.code.form</field>
-        <field name="model">l10n_tr_nilvera_einvoice_extended.account.tax.code</field>
+    <record id="l10n_tr_nilvera_einvoice_account_tax_code_view_form" model="ir.ui.view">
+        <field name="name">l10n_tr_nilvera_einvoice.account.tax.code.form</field>
+        <field name="model">l10n_tr_nilvera_einvoice.account.tax.code</field>
         <field name="arch" type="xml">
             <form string="Tax Code">
                 <div class="alert alert-info text-center" role="alert">
@@ -39,10 +39,10 @@
         </field>
     </record>
 
-    <record id="action_l10n_tr_nilvera_einvoice_extended_account_tax_code_list" model="ir.actions.act_window">
+    <record id="action_l10n_tr_nilvera_einvoice_account_tax_code_list" model="ir.actions.act_window">
         <field name="name">GIB Codes</field>
-        <field name="res_model">l10n_tr_nilvera_einvoice_extended.account.tax.code</field>
-        <field name="view_id" ref="l10n_tr_nilvera_einvoice_extended_account_tax_code_view_list"/>
+        <field name="res_model">l10n_tr_nilvera_einvoice.account.tax.code</field>
+        <field name="view_id" ref="l10n_tr_nilvera_einvoice_account_tax_code_view_list"/>
         <field name="help" type="html">
             <p class="o_view_nocontent_smiling_face">
                 Create a GIB Reason
@@ -52,7 +52,7 @@
 
     <menuitem
         id="l10n_tr_menu_account_tax_code"
-        action="action_l10n_tr_nilvera_einvoice_extended_account_tax_code_list"
+        action="action_l10n_tr_nilvera_einvoice_account_tax_code_list"
         parent="account.account_account_menu"
         sequence="9"
         groups="base.group_no_one"/>
