+++ addons/l10n_tr_nilvera_einvoice/views/product_views.xml
@@ -2,8 +2,8 @@
 <odoo>
     <data>
         <!-- view used for product.template only -->
-        <record id="product_template_only_form_view_inherit_l10n_tr_nilvera_extended" model="ir.ui.view">
-            <field name="name">product.template.product.form.view.inherit.l10n_tr_nilvera_extended</field>
+        <record id="product_template_only_form_view_inherit_l10n_tr_nilvera" model="ir.ui.view">
+            <field name="name">product.template.product.form.view.inherit.l10n_tr_nilvera</field>
             <field name="model">product.template</field>
             <field name="inherit_id" ref="product.product_template_only_form_view"/>
             <field name="arch" type="xml">
@@ -17,8 +17,8 @@
                 </field>
         </record>
         <!-- view used for product.product only -->
-        <record id="product_product_only_form_view_inherit_l10n_tr_nilvera_extended" model="ir.ui.view">
-            <field name="name">product.product.form.view.inherit.l10n_tr_nilvera_extended</field>
+        <record id="product_product_only_form_view_inherit_l10n_tr_nilvera" model="ir.ui.view">
+            <field name="name">product.product.form.view.inherit.l10n_tr_nilvera</field>
             <field name="model">product.product</field>
             <field name="inherit_id" ref="product.product_normal_form_view"/>
             <field name="arch" type="xml">
