+++ addons/l10n_tr_nilvera_einvoice/views/res_company_views.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="view_company_form_inherit_l10n_tr_nilvera_extended" model="ir.ui.view">
-        <field name="name">res.company.form.inherit.l10n_tr_nilvera_extended</field>
+    <record id="view_company_form_inherit_l10n_tr_nilvera" model="ir.ui.view">
+        <field name="name">res.company.form.inherit.l10n_tr_nilvera</field>
         <field name="model">res.company</field>
         <field name="inherit_id" ref="base.view_company_form"/>
         <field name="arch" type="xml">
