+++ addons/l10n_tr_nilvera_einvoice/views/res_config_settings_views.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-    <record id="res_config_settings_view_form_l10n_tr_nilvera_extended" model="ir.ui.view">
-        <field name="name">res.config.settings.view.form.inherit.l10n_tr_nilvera_extended</field>
+    <record id="res_config_settings_view_form_l10n_tr_nilvera" model="ir.ui.view">
+        <field name="name">res.config.settings.view.form.inherit.l10n_tr_nilvera</field>
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="l10n_tr_nilvera.res_config_settings_view_form"/>
         <field name="arch" type="xml">
