+++ addons/l10n_tr_nilvera_einvoice/views/res_partner_views.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <data>
-        <record id="view_partner_form_l10n_tr_nilvera_extended" model="ir.ui.view">
-            <field name="name">view.partner.form.inherit.l10n_tr_nilvera_extended</field>
+        <record id="view_partner_form_l10n_tr_nilvera" model="ir.ui.view">
+            <field name="name">view.partner.form.inherit.l10n_tr_nilvera</field>
             <field name="model">res.partner</field>
             <field name="inherit_id" ref="base.view_partner_form"/>
             <field name="arch" type="xml">
