--- addons/l10n_tr_nilvera_einvoice_extended/__init__.py
@@ -1,1 +1,0 @@
-from . import models
