--- addons/l10n_tr_nilvera_einvoice_extended/demo/demo_company.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="base.partner_demo_company_tr" model="res.partner">
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice_extended.l10n_tr_nilvera_einvoice_extended_tax_office_1045"></field>
-    </record>
-
-    <record id="l10n_tr.l10n_tr_res_partner_1" model="res.partner">
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice_extended.l10n_tr_nilvera_einvoice_extended_tax_office_763"></field>
-    </record>
-
-    <record id="l10n_tr.l10n_tr_res_partner_2" model="res.partner">
-        <field name="l10n_tr_tax_office_id" ref="l10n_tr_nilvera_einvoice_extended.l10n_tr_nilvera_einvoice_extended_tax_office_545"></field>
-    </record>
-</odoo>
