--- addons/l10n_tr_nilvera_einvoice_extended/security/ir.model.access.csv
@@ -1,8 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_l10n_tr_nilvera_einvoice_extended_account_tax_code_readonly,l10n_tr_nilvera_einvoice_extended_account_tax_code_readonly,model_l10n_tr_nilvera_einvoice_extended_account_tax_code,account.group_account_readonly,1,0,0,0
-access_l10n_tr_nilvera_einvoice_extended_account_tax_code,l10n_tr_nilvera_einvoice_extended_account_tax_code,model_l10n_tr_nilvera_einvoice_extended_account_tax_code,account.group_account_basic,1,1,1,1
-access_l10n_tr_nilvera_einvoice_extended_tax_office_public,l10n_tr_nilvera_einvoice_extended_tax_office_group_public_user,model_l10n_tr_nilvera_einvoice_extended_tax_office,base.group_public,1,0,0,0
-access_l10n_tr_nilvera_einvoice_extended_tax_office_portal,l10n_tr_nilvera_einvoice_extended_tax_office_group_protal_user,model_l10n_tr_nilvera_einvoice_extended_tax_office,base.group_portal,1,0,0,0
-access_l10n_tr_nilvera_einvoice_extended_tax_office_employee,l10n_tr_nilvera_einvoice_extended_tax_office_group_user,model_l10n_tr_nilvera_einvoice_extended_tax_office,base.group_user,1,0,0,0
-access_l10n_tr_nilvera_einvoice_extended_tax_office_group_partner_manager,l10n_tr_nilvera_einvoice_extended_tax_office_group_partner_manager,model_l10n_tr_nilvera_einvoice_extended_tax_office,base.group_partner_manager,1,1,1,1
-access_l10n_tr_nilvera_einvoice_extended_tax_office_group_system,l10n_tr_nilvera_einvoice_extended_tax_office_group_system,model_l10n_tr_nilvera_einvoice_extended_tax_office,base.group_system,1,1,1,1
