--- addons/l10n_tr_nilvera_einvoice_extended/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_xml_ubl_tr
