+++ odoo/addons/base/i18n/base.pot
@@ -3856,6 +3856,13 @@ msgstr ""
 msgid ""
 "\n"
 "For sending and receiving electronic invoices to Nilvera.\n"
+"\n"
+"Features include:\n"
+"\n"
+"- Support for invoice scenarios: Basic, Export, and Public Sector\n"
+"- Support for invoice types: Sales, Withholding, Tax Exempt, and Registered for Export\n"
+"- Configuration of withholding reasons and exemption reasons\n"
+"- Addition of Tax Offices.\n"
 msgstr ""
 
 #. module: base
