+++ odoo/addons/base/i18n/tr.po
@@ -9889,6 +9889,29 @@ msgstr ""
 "    "
 
 #. module: base
+#: model:ir.module.module,description:base.module_l10n_tr_nilvera_einvoice
+msgid ""
+"\n"
+"For sending and receiving electronic invoices to Nilvera.\n"
+"\n"
+"Features include:\n"
+"\n"
+"- Support for invoice scenarios: Basic, Export, and Public Sector\n"
+"- Support for invoice types: Sales, Withholding, Tax Exempt, and Registered for Export\n"
+"- Configuration of withholding reasons and exemption reasons\n"
+"- Addition of Tax Offices.\n"
+msgstr ""
+"\n"
+"Nilvera'ya elektronik fatura göndermek ve almak için.\n"
+"\n"
+"Özellikler şunları içerir:\n"
+"\n"
+"- Fatura senaryoları desteği: Temel, İhracat ve Kamu Sektörü\n"
+"- Fatura tipleri desteği: Satış, Tevkifat, İstisna ve İhraç Kayıtlı\n"
+"- Tevkifat sebepleri ve istisna sebepleri yapılandırması\n"
+"- Kontaklara Vergi Daireleri ekleme.\n"
+
+#. module: base
 #: model:ir.module.module,description:base.module_l10n_ua
 msgid ""
 "\n"
@@ -33093,6 +33116,11 @@ msgstr "Türkiye - Muhasebe"
 #: model:ir.module.module,shortdesc:base.module_l10n_tr_reports
 msgid "Türkiye - Accounting Reports"
 msgstr "Türkiye - Muhasebe Raporları"
+
+#. module: base
+#: model:ir.module.module,shortdesc:base.module_l10n_tr_nilvera_einvoice
+msgid "Türkiye - Nilvera E-Invoice"
+msgstr "Türkiye - Nilvera E-Fatura"
 
 #. module: base
 #: model:res.country,name:base.tm
